requirements.txt
google-genai
python-dotenv
flask
flake8
youtube-transcript-api
